split event test in two and move to data/tests
